Added calculating payment for new visit

+++ src/main/java/com/projectbolek/domain/model/dto/ChargeDTO.java
@@ -17,7 +17,7 @@ public class ChargeDTO extends BaseDTO implements Serializable{
 
     private Long id;
 
-    private Long toPay;
+    private Double toPay;
 
     private Timestamp invoiceDate;
